+++ contracts/test/TestPriceFeed.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity 0.7.6;
 
-import { IPriceFeed } from "../interface/IPriceFeed.sol";
+import { IPriceFeedV2 } from "../interface/IPriceFeedV2.sol";
 
 contract TestPriceFeed {
     address public chainlink;
@@ -20,30 +20,30 @@ contract TestPriceFeed {
     //
     function fetchChainlinkPrice(uint256 interval) external {
         for (uint256 i = 0; i < 17; i++) {
-            IPriceFeed(chainlink).getPrice(interval);
+            IPriceFeedV2(chainlink).getPrice(interval);
         }
-        currentPrice = IPriceFeed(chainlink).getPrice(interval);
+        currentPrice = IPriceFeedV2(chainlink).getPrice(interval);
     }
 
     function fetchBandProtocolPrice(uint256 interval) external {
         for (uint256 i = 0; i < 17; i++) {
-            IPriceFeed(bandProtocol).getPrice(interval);
+            IPriceFeedV2(bandProtocol).getPrice(interval);
         }
-        currentPrice = IPriceFeed(bandProtocol).getPrice(interval);
+        currentPrice = IPriceFeedV2(bandProtocol).getPrice(interval);
     }
 
     function cachedChainlinkPrice(uint256 interval) external {
         for (uint256 i = 0; i < 17; i++) {
-            IPriceFeed(chainlink).cacheTwap(interval);
+            IPriceFeedV2(chainlink).cacheTwap(interval);
         }
-        currentPrice = IPriceFeed(chainlink).cacheTwap(interval);
+        currentPrice = IPriceFeedV2(chainlink).cacheTwap(interval);
     }
 
     function cachedBandProtocolPrice(uint256 interval) external {
         for (uint256 i = 0; i < 17; i++) {
-            IPriceFeed(bandProtocol).cacheTwap(interval);
+            IPriceFeedV2(bandProtocol).cacheTwap(interval);
         }
-        currentPrice = IPriceFeed(bandProtocol).cacheTwap(interval);
+        currentPrice = IPriceFeedV2(bandProtocol).cacheTwap(interval);
     }
 
     //
@@ -53,7 +53,7 @@ contract TestPriceFeed {
     // having this function for testing getPrice() and cacheTwap()
     // timestamp moves if any txs happen in hardhat env and which causes cacheTwap() will recalculate all the time
     function getPrice(uint256 interval) external returns (uint256 twap, uint256 cachedTwap) {
-        twap = IPriceFeed(bandProtocol).getPrice(interval);
-        cachedTwap = IPriceFeed(bandProtocol).cacheTwap(interval);
+        twap = IPriceFeedV2(bandProtocol).getPrice(interval);
+        cachedTwap = IPriceFeedV2(bandProtocol).cacheTwap(interval);
     }
 }